requirements.txt
discord.py
asyncpg
python-dotenv
requests
aiohttp